Make some corrections

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -10,17 +10,17 @@ int main(void)
 {
 	int i;
 
-	unsigned long float sum = 1.0;
-	unsigned long float j = 2.0;
-	unsigned long float k;
+	long double sum = 1.0;
+	long double j = 2.0;
+	long double k;
 
-	printf("%.0lf, ", sum);
+	printf("%.0Lf, ", sum);
 	for (i = 1; i < 98; i++)
 	{
 		k = sum + j; /* at first run, sum =3*/
 		sum = j;
 		j = k;
-		printf("%.0lf", sum); /* use capital L for long double*/
+		printf("%.0Lf", sum); /* use capital L for long double*/
 
 		if (i != 97)
 		{
